Creates methods to add and delete interfaces in the implements list.

+++ plugins/src/main/java/com/ericliu/InterfaceMigrator.java
@@ -6,7 +6,6 @@ import com.intellij.notification.Notifications;
 import com.intellij.openapi.project.Project;
 import com.intellij.psi.JavaPsiFacade;
 import com.intellij.psi.PsiClass;
-import com.intellij.psi.PsiClassType;
 import com.intellij.psi.PsiElementFactory;
 import com.intellij.psi.PsiJavaCodeReferenceElement;
 import com.intellij.psi.PsiReferenceList;
@@ -24,6 +23,7 @@ class InterfaceMigrator {
 
     private final PsiClass boatClass;
     private final PsiClass shipClass;
+    private final PsiClass vehicleClass;
 
     InterfaceMigrator(final Project project) {
         this.project = project;
@@ -31,6 +31,7 @@ class InterfaceMigrator {
         psiElementFactory = PsiElementFactory.SERVICE.getInstance(project);
         boatClass = toPsiClass(Boat.class);
         shipClass = toPsiClass(Ship.class);
+        vehicleClass = toPsiClass(Vehicle.class);
     }
 
     void migrate() {
@@ -43,27 +44,28 @@ class InterfaceMigrator {
                 GlobalSearchScope.allScope(project)).findAll()) {
 
             final PsiReferenceList implementsList = boatImplementor.getImplementsList();
-            removeBoatInterface(implementsList);
 
-            addShipInterface(boatImplementor, implementsList);
+            addInterface(vehicleClass, implementsList);
+
+            removeInterface(boatClass, implementsList);
+
+            addInterface(shipClass, implementsList);
         }
     }
 
-    private void addShipInterface(final PsiClass boatImplementor, final PsiReferenceList implementsList) {
-        implementsList.add(psiElementFactory.createClassReferenceElement(shipClass));
+    private void addInterface(final PsiClass interfaceClass, final PsiReferenceList implementsList) {
+        implementsList.add(psiElementFactory.createClassReferenceElement(interfaceClass));
 
         Notifications.Bus.notify(new Notification(InterfaceMigrator.class.getName(),
-                "Found Implementor: " + boatImplementor.getName(),
-                "Adds interface " + shipClass.getName(),
+                "Found Implementor",
+                "Adds interface " + interfaceClass.getName(),
                 NotificationType.INFORMATION));
     }
 
-    private void removeBoatInterface(final PsiReferenceList implementsList) {
-        final PsiClassType boatType = psiElementFactory.createType(boatClass);
-
+    private void removeInterface(final PsiClass interfaceClass, final PsiReferenceList implementsList) {
         Notifications.Bus.notify(new Notification(InterfaceMigrator.class.getName(),
-                "Removing boat",
-                "Reference name: " + boatType.getClassName(),
+                "Removing",
+                "Interface name: " + interfaceClass.getName(),
                 NotificationType.INFORMATION));
 
         for (final PsiJavaCodeReferenceElement referenceElement : implementsList.getReferenceElements()) {
@@ -72,7 +74,14 @@ class InterfaceMigrator {
                     "Reference name: " + referenceElement.getQualifiedName(),
                     NotificationType.INFORMATION));
 
-            referenceElement.delete();
+            if (referenceElement.getQualifiedName().equals(interfaceClass.getQualifiedName())) {
+
+                Notifications.Bus.notify(new Notification(InterfaceMigrator.class.getName(),
+                        "Found interface to remove",
+                        referenceElement.getQualifiedName(),
+                        NotificationType.INFORMATION));
+                referenceElement.delete();
+            }
         }
     }
